Move to using dirs for different manufacturers
