+++ skrf/vi/vna/__init__.py
@@ -4,13 +4,6 @@
 Vector Network Analyzers (:mod:`skrf.vi.vna`)
 ++++++++++++++++++++++++++++++++++++++++++++++++++++
 
-New VNA drivers
----------------
-
-- VNA drivers will now have a common high level functionality across all vendors implemented in an ABCVNA class.
-- Different vendor drivers will implement their own mid level functionality as needed to implement the ABC class
-- The low level functions are all implemented as SCPI commands which have a new way of being generated and called
-
 Available VNAs
 ------------------
 
@@ -18,11 +11,8 @@ Available VNAs
     :toctree: generated/
 
     PNA
-    ZVA40
-    HP8510C
-    HP8720
-    NanoVNAv2
 """
 
-from .pna import PNA
+
+from .keysight.pna import PNA
 from .vna import VNA
